Update docstring in step_function.py

+++ src/pymovements/synthetic/step_function.py
@@ -49,7 +49,7 @@ def step_function(
         If greater than zero, gaussian noise is scaled according to value and superimposed on the
         output array. (default: 0)
     dtype: np.typing.DTypeLike
-        The desired data type of the output array. (default: np.float32)
+        The desired data type of the output array. (default: np.float64)
 
     Returns
     -------
@@ -70,7 +70,7 @@ def step_function(
     ...     values=[1., 2., 3.],
     ...     start_value=0,
     ... )
-    array([0., 0., 1., 1., 1., 2., 2., 2., 2., 3.], dtype=float32)
+    array([0., 0., 1., 1., 1., 2., 2., 2., 2., 3.])
 
     >>> # multi-channel example
     >>> step_function(
@@ -88,7 +88,7 @@ def step_function(
            [ 3.,  4.],
            [ 3.,  4.],
            [ 3.,  4.],
-           [ 3.,  4.]], dtype=float32)
+           [ 3.,  4.]])
     """
     # Check that steps and values have equal length.
     if len(steps) != len(values):
